LayoutTests/imported/w3c/web-platform-tests/mathml/relations/css-styling/lengths-2.html

<!DOCTYPE html><!-- webkit-test-runner [ CoreMathMLEnabled=true ] -->
<html>
<head>
<meta charset="utf-8">
<title>MathML lengths</title>
<link rel="help" href="https://w3c.github.io/mathml-core/#css-styling">
<link rel="help" href="https://w3c.github.io/mathml-core/#types-for-mathml-attribute-values">
<link rel="help" href="https://w3c.github.io/mathml-core/#legacy-mathml-style-attributes">
<link rel="help" href="https://w3c.github.io/mathml-core/#space-mspace">
<meta name="assert" content="Verify various cases of the MathML length syntax.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/mathml/support/feature-detection.js"></script>
<script src="/mathml/support/fonts.js"></script>
<style>
  @font-face {
    font-family: TestFont;
    src: url("/fonts/math/xheight500.woff");
  }
  math {
    font-family: TestFont;
    font-size: 10px;
  }
</style>
<script>
  var epsilon = .5;

  function getBox(aId) {
    return document.getElementById(aId).getBoundingClientRect();
  }

  setup({ explicit_done: true });
  window.addEventListener("load", () => { loadAllFonts().then(runTests); });

  function runTests() {
    test(function() {
      assert_equals(getBox("unitEx").width, 500, "ex");
    }, "Units");

    done();
  }
</script>
</head>
<body>
  <div id="log"></div>
  <p>
    <math>
      <mspace id="unitEx" width="100ex"/>
    </math>
  </p>
</body>
</html>
